+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3 (development version)
+
 # mlr3 1.7.0
 
 * BREAKING CHANGE: Removed the deprecated `Task$divide()` method, use the `$internal_valid_task` field instead.